Initialize PokeMe project
Initial set up of the `PokeMe` project.
This project is a proxy of `PokeAPI` for some of the object defined in
this API.

It is used only as an exercise.

The project uses docker, as this will ease the setup, needed to run it.

We include also a psql container to use for persistence.